* Fixed a few possible data loss warnings.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@5372 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ combat.cpp
@@ -1240,10 +1240,10 @@ void AreaCombat::copyArea(const MatrixArea* input, MatrixArea* output, MatrixOpe
 		}
 		double angleRad = 3.1416 * angle / 180.0;
 
-		float a = std::cos(angleRad);
-		float b = -std::sin(angleRad);
-		float c = std::sin(angleRad);
-		float d = std::cos(angleRad);
+		double a = std::cos(angleRad);
+		double b = -std::sin(angleRad);
+		double c = std::sin(angleRad);
+		double d = std::cos(angleRad);
 
 		for(int32_t x = 0; x < (long)input->getCols(); ++x)
 		{
